ci: make release workflow the only release path
